refactor : LandingSprint 리팩토링 및 코드 수정
1. LandingSprint 컴포넌트를 분리
2. diffBetweenDate 함수의 경우, 동일한 날짜이더라도 100%가 뜨지 않는 현상이 발생
- 이를 해결하기 위해, 시간을 제외한 날짜단위로 계산하도록 코드를 수정
